# server/docs/swagger.yaml
definitions:
  cms.Applicant:
    properties:
      ID:
        description: 主键ID
        type: integer
      code:
        description: 统一社会信用代码
        type: string
      company:
        description: 公司名
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - code
    - company
    type: object
...
  cms.Court:
    properties:
      ID:
        description: 主键ID
        type: integer
      addr:
        description: 地址
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      name:
        description: 法院名
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - addr
    - name
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: cms-gva Swagger API接口文档
  version: v0.0.1
paths:
...
  /applicant/createApplicant:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建申请人
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Applicant'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建申请人
      tags:
      - Applicant
  /applicant/deleteApplicant:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除申请人
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Applicant'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除申请人
      tags:
      - Applicant
  /applicant/deleteApplicantByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除申请人
      tags:
      - Applicant
  /applicant/findApplicant:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 统一社会信用代码
        in: query
        name: code
        required: true
        type: string
      - description: 公司名
        in: query
        name: company
        required: true
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    reapplicant:
                      $ref: '#/definitions/cms.Applicant'
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询申请人
      tags:
      - Applicant
  /applicant/getApplicantList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取申请人列表
      tags:
      - Applicant
  /applicant/getApplicantPublic:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 不需要鉴权的申请人接口
      tags:
      - Applicant
  /applicant/updateApplicant:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新申请人
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Applicant'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新申请人
      tags:
      - Applicant
...
  /court/createCourt:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建法院
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Court'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建法院
      tags:
      - Court
  /court/deleteCourt:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除法院
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Court'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除法院
      tags:
      - Court
  /court/deleteCourtByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除法院
      tags:
      - Court
  /court/findCourt:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 地址
        in: query
        name: addr
        required: true
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - description: 法院名
        in: query
        name: name
        required: true
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    recourt:
                      $ref: '#/definitions/cms.Court'
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询法院
      tags:
      - Court
  /court/getCourtList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取法院列表
      tags:
      - Court
  /court/getCourtPublic:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 不需要鉴权的法院接口
      tags:
      - Court
  /court/updateCourt:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新法院
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cms.Court'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新法院
      tags:
      - Court
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"